% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond_plot.R
\name{diamond_plot}
\alias{diamond_plot}
\title{diamond plot}
\usage{
diamond_plot(df, varx, vary)
}
\arguments{
\item{df}{a dataframe}

\item{varx}{a string name for variable x in the dataframe df}

\item{vary}{a string name for variable y in the dataframe df}
}
\value{
A plot with diamond markers of {vary} versus {varx}
}
\description{
This function produces a ggplot plot of {vary} versus {varx}
}
\details{
with diamond markers.
}
\examples{
df.test <- data.frame(x = rnorm(10),
                      y = rnorm(10))
diamond_plot(df.test, 'x', 'y')

}
